requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
bcrypt>=4.0.0
PyYAML>=6.0
requests>=2.31.0
Flask>=2.3.2
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.5
werkzeug>=2.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
imutils>=0.5.4
spacy>=3.7.0
easyocr>=1.7.0
